Improve video verification by adding type safety for trust tier calculations
Add explicit integer type casts to parameters used in trust tier calculations within the handleSaveVideo and handleVerifyVideo functions to resolve PostgreSQL type deduction errors.

+++ api/index.ts
@@ -1649,8 +1649,8 @@ async function handleSaveVideo(req: VercelRequest, res: VercelResponse) {
         `UPDATE students 
          SET video_approval_streak = COALESCE(video_approval_streak, 0) + 1,
              trust_tier = CASE 
-               WHEN COALESCE(video_approval_streak, 0) + 1 >= $1 THEN 'trusted'
-               WHEN COALESCE(video_approval_streak, 0) + 1 >= $2 THEN 'verified'
+               WHEN COALESCE(video_approval_streak, 0) + 1 >= $1::integer THEN 'trusted'
+               WHEN COALESCE(video_approval_streak, 0) + 1 >= $2::integer THEN 'verified'
                ELSE COALESCE(trust_tier, 'unverified')
              END,
              updated_at = NOW()
@@ -1885,8 +1885,8 @@ async function handleVerifyVideo(req: VercelRequest, res: VercelResponse, videoI
         `UPDATE students 
          SET video_approval_streak = COALESCE(video_approval_streak, 0) + 1,
              trust_tier = CASE 
-               WHEN COALESCE(video_approval_streak, 0) + 1 >= $1 THEN 'trusted'
-               WHEN COALESCE(video_approval_streak, 0) + 1 >= $2 THEN 'verified'
+               WHEN COALESCE(video_approval_streak, 0) + 1 >= $1::integer THEN 'trusted'
+               WHEN COALESCE(video_approval_streak, 0) + 1 >= $2::integer THEN 'verified'
                ELSE COALESCE(trust_tier, 'unverified')
              END,
              updated_at = NOW()
